error on create certificate is not handled

+++ assets/cert-creation.sh
@@ -1,4 +1,6 @@
 #!/bin/sh
+set -e
+
 domain=$1
 email=$2
 
@@ -7,7 +9,8 @@ folder="/etc/letsencrypt/live/$domain"
 if [ ! -d $folder ]; then
     certbot certonly --text -n --keep --email $email --agree-tos --webroot -w /app/letsencrypt -d $domain
     #    --staging
-
-    cat ${folder}/privkey.pem ${folder}/fullchain.pem | tee ${folder}/haproxy.pem >/dev/null
+    if [ -e ${folder}/fullchain.pem ]; then
+        cat ${folder}/privkey.pem ${folder}/fullchain.pem | tee ${folder}/haproxy.pem >/dev/null
+    fi
 fi
